LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/nested-browsing-contexts/window-parent-null-expected.txt


PASS `window.parent` is null when browsing context container element removed 
PASS `window.parent` null when parent browsing context container removed 
 
